RiscV: Only init OpenSBI once, in SEC
OpenSBI can only be initialized once. Therefore we do it as early as possible -
in SEC, while switching to PEI.
To switch to DXE later from PEI we call sbi_hart_switch_mode. This function is saved in the firmware specific context.
Because we don't have access to the OpenSBI library later in PEI and DXE we create the PPI to call OpenSBI in SEC.